Resolves #3 - UtkScanInterface::Setup() return value ignored
This was more complicated than it looked. I needed to propagate the
scan_init variable to the children classes. Then I needed to update the
UtkUnpacker destructor. The destructor was instancing DetectorDriver before
trying to destruct it! This means that it was actually constructing
DetectorDriver just so that it could destruct it. I added a bool to store
the status of the DetectorDriver. Finally, I put all the calls except
UtkScanInterface::close inside the try/catch block.