Resolves #28 - Get ROOT working in the CI Server
I figured out how to install root on the CI server so that we can compile
successfully against the libraries. It's ROOT5 but that's better than
nothing. The PAASS_BUILD_SCAN_UTILITIES=OFF because there are issues
building rootScanner.

I consider this good enough for now since rootScanner is still in pre-beta.