Resolves #110 - Build fails without firmware installed
Technically we don't have to have firmware to build the software. We just
need to have the API. I've updated the build to provide an error message
for the user in pixie.cfg. There's also a status message that's output during the
build.

I modified the SUBDIRLIST logic to check if we have any elements in dirlist before
we try to remove duplicates. The REMOVE_DUPLICATES command requires that there be
at least one element in the list.