Updates doxygen documentation
I moved the doxyfile to the root directory. This prevents people from accidentally
running doxygen from the Doc directory. I've removed all the commenting to make the
file easier to read and edit. All the information for the options is on the doxygen
website. I removed the images; the wiki has them anyway. Doxygen creates the Doc folder
during compilation.

I removed the main page, and replaced it with the README.md. Now we only have to updated
that information once.