Cleaning up unittest and cd()s
We had many more cd() calls than needed. I replaced these by associating the objects
with the proper directory. We didn't get away from the cd() while writing the
histograms to disk. Not sure why we needed that one still.

I updated the fill call to only try filling histograms that have a non-zero number of
entries. I don't have any evidence that this is more efficient. I think it makes sense
that we don't try to write a histogram that has zero entries though.

I updated the unittest to check the returned histogram class. This will control any potential
changes in the logic assigning the histograms. This also allows us to cut out the part that
filled histograms so that we could manually check their assigned classes.