Incorporating new interface classes into poll2. Committing interfaces.
We broke up the interface classes into the following structure :
AcquisitionInterface
 |--- EmulatedInterface
 |--- PixieInterface

We can interchange the classes at this point to test components of the software.
Poll2 still uses only PixieInterface since this is a more complex change than the other
programs. I updated poll2 to use the latest version of the interface methods and classes.

The emulated interface is MVP complete, but still needs some additional tweaks so that programs
will run correctly.