Converting pixie.cfg and slot_def.set to a single XML document
Moved reading pixie.cfg and slot_def.set to the same place. No reason that they had to
be broken up.

We now include a crap ton of error handling. This gives users a much better idea of what's
going on, why things failed, and what they need to do to fix it.

VerifyFiles provides a way for users to verify that all the files exist before booting
the modules. I didn't make this automatic so that I could unit test reading the XML.

Removed the definition of the number of modules from the top level of slot_def.set.
There was no reason to define that and then define a list of modules. Just define what you
need and nothing else.