Removes the MCA_args class
This class could be completely replaced with just a couple of bools and double. It served no real purpose. 

Updated the mca command parsing so that it's more robust, and allows users to use some default arguments. 

Reworked the MCA run control so that we could run without mca_args class.