EmulatedInterface now overrides PixieInterface all poll2 cmds work
The first iteration of EmulatedInterface is complete! The software can write data files to disk!!!

The scan code unpacks correctly.  We'll need to setup the event time low and event time high to increase instead of be a constant value. 

Made updates to formatting in PixieSupport to improve readability. Still need to add documentation and clean up style.